requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.8.0
python-engineio==4.7.1
Werkzeug==2.3.7
pytest==7.4.3
pytest-flask==1.3.0